[refactor] diet app rename
