+++ diet/admin.py
