+++ diet/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class DietsConfig(AppConfig):
+class DietConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'diets'
+    name = 'diet'
