+++ eatda_be/urls.py
@@ -22,7 +22,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('accounts/', include('account.urls')),
     path('users/', include('users.urls')),
-    path('diets/', include('diets.urls'))
+    path('diets/', include('diet.urls'))
 ]
 
 # 개발 환경 미디어 세팅
