+++ users/models.py
@@ -2,7 +2,7 @@ from django.core.validators import MinValueValidator, MaxValueValidator
 from account.models import BaseModel
 from django.db import models
 from django.contrib.auth import get_user_model
-from diets.models import Data, Allergy
+from diet.models import Data, Allergy
 
 User = get_user_model()
